refactor : User 엔티티에서 OAuth2User 분리
User 엔티티 대신 UserPrincipal 클래스를 새로 생성하여,
OAuth2User implement 하였습니다.
위 변경에 따른 KakaoOAuth2UserService 코드도 수정하였습니다.
