+++ src/main/java/leejuyeon/pacemaker/entity/User.java
@@ -1,19 +1,19 @@
 package leejuyeon.pacemaker.entity;
 
-import jakarta.persistence.*;
-import leejuyeon.pacemaker.enums.Authority;
-import lombok.*;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import java.time.LocalDateTime;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-
-import java.time.LocalDateTime;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Map;
 
 @EntityListeners(AuditingEntityListener.class)
 @AllArgsConstructor
@@ -22,48 +22,30 @@ import java.util.Map;
 @Setter
 @Getter
 @Entity
-public class User implements OAuth2User {
+public class User {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
 
-    private String email;
+  private String email;
 
-    private String nickname;
+  private String nickname;
 
-    private Float weight;
+  private Float weight;
 
-    private Float muscleMass;
+  private Float muscleMass;
 
-    private Float fatMass;
+  private Float fatMass;
 
-    private boolean visibility;
+  private boolean visibility;
 
-    private Long totalLikes;
+  private Long totalLikes;
 
-    @Enumerated(EnumType.STRING)
-    private Authority authority;
+  @CreatedDate
+  private LocalDateTime createdAt;
 
-    @CreatedDate
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
-
-    @Override
-    public Map<String, Object> getAttributes() {
-        return null;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return Collections.singleton(new SimpleGrantedAuthority(authority.getRole()));
-    }
-
-    @Override
-    public String getName() {
-        return email;
-    }
+  @LastModifiedDate
+  private LocalDateTime updatedAt;
 
 }
